Added new check type for email round trip and file exists check
The first of these two new checks will send an email out via SMTP and
await a response back from a an IMAP mail box.  The roundtrip is also
timed.
The second looks for a file to exist at a specific location.  The name
of the file can be built using a c# expression to allow these names to
be dynamically built based on stuff like the date/time.